Implemented BadIdExceptionFilter: - Implemented tests; - Implemented filters for /GET, /PATCH, /PUT, /DELETE methods; - Implemented logic for returning BadRequest response for not a number URL :id

+++ backend/src/todo/todo.controller.ts
@@ -6,10 +6,14 @@ import {
   Patch,
   Param,
   Delete,
+  UseFilters,
+  BadRequestException,
+  Put,
 } from '@nestjs/common';
 import { TodoService } from './todo.service';
 import { CreateTodoDto } from './dto/create-todo.dto';
 import { UpdateTodoDto } from './dto/update-todo.dto';
+import { BadIdExceptionFilter } from 'src/badid/bad-id.filter';
 
 @Controller('todo')
 export class TodoController {
@@ -26,17 +30,30 @@ export class TodoController {
   }
 
   @Get(':id')
+  @UseFilters(BadIdExceptionFilter)
   findOne(@Param('id') id: string) {
+    if (isNaN(+id)) {
+      throw new BadRequestException();
+    }
     return this.todoService.findOne(+id);
   }
 
+  @Put(':id')
   @Patch(':id')
+  @UseFilters(BadIdExceptionFilter)
   update(@Param('id') id: string, @Body() updateTodoDto: UpdateTodoDto) {
+    if (isNaN(+id)) {
+      throw new BadRequestException();
+    }
     return this.todoService.update(+id, updateTodoDto);
   }
 
   @Delete(':id')
+  @UseFilters(BadIdExceptionFilter)
   remove(@Param('id') id: string) {
+    if (isNaN(+id)) {
+      throw new BadRequestException();
+    }
     return this.todoService.remove(+id);
   }
 }
